Split netcoreapp code to its own files
