+++ src/System.Runtime/tests/Helpers.cs
@@ -8,30 +8,8 @@ using System.Reflection.Emit;
 
 namespace System.Tests
 {
-    public static class Helpers
+    public static partial class Helpers
     {
-#if netcoreapp // AssemblyBuilder is not in netstandard yet 
-
-        private static Type s_nonRuntimeType;
-
-        public static Type NonRuntimeType()
-        {
-            if (s_nonRuntimeType == null)
-            {
-                AssemblyName assemblyName = new AssemblyName("AssemblyName");
-                AssemblyBuilder assemblyBuilder = AssemblyBuilder.DefineDynamicAssembly(assemblyName, AssemblyBuilderAccess.Run);
-                ModuleBuilder mboduleBuilder = assemblyBuilder.DefineDynamicModule(assemblyName.Name);
-
-                TypeBuilder typeBuilder = mboduleBuilder.DefineType("TestType", TypeAttributes.Public);
-
-                GenericTypeParameterBuilder[] typeParams = typeBuilder.DefineGenericParameters("T");
-                s_nonRuntimeType = typeParams[0].UnderlyingSystemType;
-            }
-
-            return s_nonRuntimeType;
-        }
-#endif // netcoreapp
-
         public static void PerformActionWithCulture(CultureInfo culture, Action test)
         {
             CultureInfo originalCulture = CultureInfo.CurrentCulture;
