+++ src/System.Runtime/tests/System.Runtime.Tests.csproj
@@ -120,6 +120,9 @@
     </Compile>
   </ItemGroup>
   <ItemGroup Condition="'$(TargetGroup)' == 'netcoreapp'">
+    <Compile Include="ArrayTests.netcoreapp.cs" />
+    <Compile Include="Helpers.netcoreapp.cs" />
+    <Compile Include="System\ActivatorTests.netcoreapp.cs" />
     <Compile Include="System\ArrayTests.netcoreapp.cs" />
     <Compile Include="System\EnumTests.netcoreapp.cs" />
     <Compile Include="System\GCTests.netcoreapp.cs" />
