+++ src/System.Runtime/tests/System/ArrayTests.cs
@@ -1431,16 +1431,6 @@ namespace System.Tests
             Assert.Throws<ArgumentOutOfRangeException>("lengths[0]", () => Array.CreateInstance(typeof(int), new int[] { -1 }, new int[1]));
         }
 
-#if netcoreapp
-        [Fact]
-        public static void CreateInstance_TypeNotRuntimeType_ThrowsArgumentException()
-        {
-            Assert.Throws<ArgumentException>("elementType", () => Array.CreateInstance(Helpers.NonRuntimeType(), 0));
-            Assert.Throws<ArgumentException>("elementType", () => Array.CreateInstance(Helpers.NonRuntimeType(), new int[1]));
-            Assert.Throws<ArgumentException>("elementType", () => Array.CreateInstance(Helpers.NonRuntimeType(), new int[1], new int[1]));
-        }
-#endif // netcoreapp
-
         [Fact]
         public static void CreateInstance_LengthsNull_ThrowsArgumentNullException()
         {
