+++ src/System.Runtime/tests/System/EnumTests.cs
@@ -334,23 +334,6 @@ namespace System.Tests
             Assert.True(s == "Green" || s == "Green_a" || s == "Green_b");
         }
 
-#if netcoreapp
-        public static IEnumerable<object[]> UnsupportedEnumType_TestData()
-        {
-            yield return new object[] { s_floatEnumType, 1.0f };
-            yield return new object[] { s_doubleEnumType, 1.0 };
-            yield return new object[] { s_intPtrEnumType, (IntPtr)1 };
-            yield return new object[] { s_uintPtrEnumType, (UIntPtr)1 };
-        }
-
-        [Theory]
-        [MemberData(nameof(UnsupportedEnumType_TestData))]
-        public static void GetName_Unsupported_ThrowsArgumentException(Type enumType, object value)
-        {
-            Assert.Throws<ArgumentException>("value", () => Enum.GetName(enumType, value));
-        }
-#endif // netcoreapp
-
         [Fact]
         public static void GetName_Invalid()
         {
@@ -478,18 +461,6 @@ namespace System.Tests
         {
             Assert.Equal(expected, Enum.IsDefined(enumType, value));
         }
-
-#if netcoreapp
-        [Theory]
-        [MemberData(nameof(UnsupportedEnumType_TestData))]
-        public static void IsDefined_UnsupportedEnumType_ThrowsInvalidOperationException(Type enumType, object value)
-        {
-            // A Contract.Assert(false, "...") is hit for certain unsupported primitives
-            Exception ex = Assert.ThrowsAny<Exception>(() => Enum.IsDefined(enumType, value));
-            string exName = ex.GetType().Name;
-            Assert.True(exName == nameof(InvalidOperationException) || exName == "ContractException");
-        }
-#endif // netcoreapp
 
         [Fact]
         public static void IsDefined_Invalid()
@@ -1484,18 +1455,6 @@ namespace System.Tests
             Format(e.GetType(), e, format, expected);
         }
 
-#if netcoreapp 
-        [Fact]
-        public static void ToString_InvalidUnicodeChars()
-        {
-            // TODO: move into ToString_Format_TestData when dotnet/buildtools#1091 is fixed
-            ToString_Format((Enum)Enum.ToObject(s_charEnumType, char.MaxValue), "D", char.MaxValue.ToString());
-            ToString_Format((Enum)Enum.ToObject(s_charEnumType, char.MaxValue), "X", "FFFF");
-            ToString_Format((Enum)Enum.ToObject(s_charEnumType, char.MaxValue), "F", char.MaxValue.ToString());
-            ToString_Format((Enum)Enum.ToObject(s_charEnumType, char.MaxValue), "G", char.MaxValue.ToString());
-        }
-#endif // netcoreapp
-
         [Fact]
         public static void ToString_Format_MultipleMatches()
         {
@@ -1505,26 +1464,6 @@ namespace System.Tests
             s = ((SimpleEnum)3).ToString("G");
             Assert.True(s == "Green" || s == "Green_a" || s == "Green_b");
         }
-
-#if netcoreapp 
-        public static IEnumerable<object[]> UnsupportedEnum_TestData()
-        {
-            yield return new object[] { Enum.ToObject(s_floatEnumType, 1) };
-            yield return new object[] { Enum.ToObject(s_doubleEnumType, 2) };
-            yield return new object[] { Enum.ToObject(s_intPtrEnumType, 1) };
-            yield return new object[] { Enum.ToObject(s_uintPtrEnumType, 2) };
-        }
-
-        [Theory]
-        [MemberData(nameof(UnsupportedEnum_TestData))]
-        public static void ToString_UnsupportedEnumType_ThrowsArgumentException(Enum e)
-        {
-            // A Contract.Assert(false, "...") is hit for certain unsupported primitives
-            Exception formatXException = Assert.ThrowsAny<Exception>(() => e.ToString("X"));
-            string formatXExceptionName = formatXException.GetType().Name;
-            Assert.True(formatXExceptionName == nameof(InvalidOperationException) || formatXExceptionName == "ContractException");
-        }
-#endif // netcoreapp
 
         [Fact]
         public static void ToString_InvalidFormat_ThrowsFormatException()
@@ -1557,26 +1496,6 @@ namespace System.Tests
             Assert.Equal(expected, Enum.Format(enumType, value, format.ToLowerInvariant()));
         }
 
-#if netcoreapp
-        [Theory]
-        [MemberData(nameof(UnsupportedEnumType_TestData))]
-        public static void Format_UnsupportedEnumType_ThrowsArgumentException(Type enumType, object value)
-        {
-            // A Contract.Assert(false, "...") is hit for certain unsupported primitives
-            Exception formatGException = Assert.ThrowsAny<Exception>(() => Enum.Format(enumType, value, "G"));
-            string formatGExceptionName = formatGException.GetType().Name;
-            Assert.True(formatGExceptionName == nameof(InvalidOperationException) || formatGExceptionName == "ContractException");
-
-            Exception formatXException = Assert.ThrowsAny<Exception>(() => Enum.Format(enumType, value, "X"));
-            string formatXExceptionName = formatXException.GetType().Name;
-            Assert.True(formatXExceptionName == nameof(InvalidOperationException) || formatXExceptionName == "ContractException");
-
-            Exception formatFException = Assert.ThrowsAny<Exception>(() => Enum.Format(enumType, value, "F"));
-            string formatFExceptionName = formatFException.GetType().Name;
-            Assert.True(formatFExceptionName == nameof(InvalidOperationException) || formatFExceptionName == "ContractException");
-        }
-#endif // netcoreapp
-
         [Fact]
         public static void Format_Invalid()
         {
@@ -1595,95 +1514,5 @@ namespace System.Tests
             Assert.Throws<FormatException>(() => Enum.Format(typeof(SimpleEnum), SimpleEnum.Red, "   \t")); // Format is whitespace
             Assert.Throws<FormatException>(() => Enum.Format(typeof(SimpleEnum), SimpleEnum.Red, "t")); // No such format
         }
-
-#if netcoreapp // EnumBuilder is defined in netcoreapp
-        private static EnumBuilder GetNonRuntimeEnumTypeBuilder(Type underlyingType)
-        {
-            AssemblyBuilder assembly = AssemblyBuilder.DefineDynamicAssembly(new AssemblyName("Name"), AssemblyBuilderAccess.Run);
-            ModuleBuilder module = assembly.DefineDynamicModule("Name");
-
-            return module.DefineEnum("TestName_" + underlyingType.Name, TypeAttributes.Public, underlyingType);
-        }
-
-        private static Type s_boolEnumType = GetBoolEnumType();
-        private static Type GetBoolEnumType()
-        {
-            EnumBuilder enumBuilder = GetNonRuntimeEnumTypeBuilder(typeof(bool));
-            enumBuilder.DefineLiteral("Value1", true);
-            enumBuilder.DefineLiteral("Value2", false);
-
-            return enumBuilder.CreateTypeInfo().AsType();
-        }
-
-        private static Type s_charEnumType = GetCharEnumType();
-        private static Type GetCharEnumType()
-        {
-            EnumBuilder enumBuilder = GetNonRuntimeEnumTypeBuilder(typeof(char));
-            enumBuilder.DefineLiteral("Value1", (char)1);
-            enumBuilder.DefineLiteral("Value2", (char)2);
-
-            enumBuilder.DefineLiteral("Value0x3f06", (char)0x3f06);
-            enumBuilder.DefineLiteral("Value0x3000", (char)0x3000);
-            enumBuilder.DefineLiteral("Value0x0f06", (char)0x0f06);
-            enumBuilder.DefineLiteral("Value0x1000", (char)0x1000);
-            enumBuilder.DefineLiteral("Value0x0000", (char)0x0000);
-            enumBuilder.DefineLiteral("Value0x0010", (char)0x0010);
-            enumBuilder.DefineLiteral("Value0x3f16", (char)0x3f16);
-
-            return enumBuilder.CreateTypeInfo().AsType();
-        }
-
-        private static Type s_floatEnumType = GetFloatEnumType();
-        private static Type GetFloatEnumType()
-        {
-            EnumBuilder enumBuilder = GetNonRuntimeEnumTypeBuilder(typeof(float));
-            enumBuilder.DefineLiteral("Value1", 1.0f);
-            enumBuilder.DefineLiteral("Value2", 2.0f);
-
-            enumBuilder.DefineLiteral("Value0x3f06", (float)0x3f06);
-            enumBuilder.DefineLiteral("Value0x3000", (float)0x3000);
-            enumBuilder.DefineLiteral("Value0x0f06", (float)0x0f06);
-            enumBuilder.DefineLiteral("Value0x1000", (float)0x1000);
-            enumBuilder.DefineLiteral("Value0x0000", (float)0x0000);
-            enumBuilder.DefineLiteral("Value0x0010", (float)0x0010);
-            enumBuilder.DefineLiteral("Value0x3f16", (float)0x3f16);
-
-            return enumBuilder.CreateTypeInfo().AsType();
-        }
-
-        private static Type s_doubleEnumType = GetDoubleEnumType();
-        private static Type GetDoubleEnumType()
-        {
-            EnumBuilder enumBuilder = GetNonRuntimeEnumTypeBuilder(typeof(double));
-            enumBuilder.DefineLiteral("Value1", 1.0);
-            enumBuilder.DefineLiteral("Value2", 2.0);
-
-            enumBuilder.DefineLiteral("Value0x3f06", (double)0x3f06);
-            enumBuilder.DefineLiteral("Value0x3000", (double)0x3000);
-            enumBuilder.DefineLiteral("Value0x0f06", (double)0x0f06);
-            enumBuilder.DefineLiteral("Value0x1000", (double)0x1000);
-            enumBuilder.DefineLiteral("Value0x0000", (double)0x0000);
-            enumBuilder.DefineLiteral("Value0x0010", (double)0x0010);
-            enumBuilder.DefineLiteral("Value0x3f16", (double)0x3f16);
-
-            return enumBuilder.CreateTypeInfo().AsType();
-        }
-
-        private static Type s_intPtrEnumType = GetIntPtrEnumType();
-        private static Type GetIntPtrEnumType()
-        {
-            EnumBuilder enumBuilder = GetNonRuntimeEnumTypeBuilder(typeof(IntPtr));
-
-            return enumBuilder.CreateTypeInfo().AsType();
-        }
-
-        private static Type s_uintPtrEnumType = GetUIntPtrEnumType();
-        private static Type GetUIntPtrEnumType()
-        {
-            EnumBuilder enumBuilder = GetNonRuntimeEnumTypeBuilder(typeof(UIntPtr));
-
-            return enumBuilder.CreateTypeInfo().AsType();
-        }
-#endif // netcoreapp
     }
 }
